Siirrettiin kuvaluokat pois gtk:n puolelta
