+++ gtk_nn_parse.py
@@ -1,9 +1,6 @@
 #!/usr/bin/env python3
 
-import os
-import requests
 import math
-import re
 import gi
 gi.require_version('Gtk', '3.0')
 from gi.repository import Gtk,Gdk
@@ -25,7 +22,7 @@ class KommenttiLaatikko(Gtk.ListBoxRow):
 		builder = Gtk.Builder()
 		builder.add_from_file("KommenttiLaatikko.glade")
 		user_thumbnail = builder.get_object("user_thumbnail")
-		user_thumbnail.set_from_file(Avatar(kommentti.user.avatar).get_file())
+		user_thumbnail.set_from_file(kommentti.user.avatar.get_file())
 		user_name = builder.get_object("user_name")
 		user_name.set_label("<"+kommentti.user.name+">")
 		user_data = builder.get_object("user_data")
@@ -147,14 +144,7 @@ class VLCWindow(Gtk.Window):
 		self.change_video("p")
 		
 	def change_video(self, direction):
-		url = "https://naurunappula.com/go.php"
-		payload = {
-			'link_id': self.data.id,
-			'c': '2',
-			'dir': direction
-		}
-		sessio = mie.get(url, params=payload)
-		self.data = VideoPage(sessio)
+		self.data = VideoPage(self.data.change_media(direction))
 		self.draw_area.player.stop()
 		self.draw_area.player.set_mrl(self.data.url)
 		self.draw_area.player.play()
@@ -163,49 +153,13 @@ class VLCWindow(Gtk.Window):
 		self.queue_draw()
 		self.show_all()
 
-class Kuva(object):
-	folder = None
-
-	def __init__(self, url):
-		self.imgname = re.search('\d+[.]jpg', url)
-		self.url = url
-		
-	def get_file(self):
-		return self.folder + self.imgname
-		
-	def write_file(self):
-		if not os.path.isfile(self.folder+self.imgname):				
-			response=requests.get(self.url)
-			with open(self.folder+self.imgname, 'wb') as img:
-				for chunk in response:
-					img.write(chunk)
-		
-class Avatar(Kuva):
-	folder = "./avatars/"
-	
-	def __init__(self, url):
-		Kuva.__init__(self, url)
-		self.imgname = self.imgname.group(0)
-		self.write_file()
-
-class Thumbnail(Kuva):
-	folder = "./thumbnails/"
-	
-	def __init__(self, url):
-		Kuva.__init__(self, url)
-		if self.imgname == None:
-			self.imgname="video.gif"
-		else:
-			self.imgname=self.imgname.group(0)
-		self.write_file()
-
 class Nappi(Gtk.Button):
 	def __init__(self, olio):
 		Gtk.Button.__init__(self)
 		self.Mnemonic = olio
 		jako = Gtk.VBox()
 		kuva = Gtk.Image()
-		kuva.set_from_file(Thumbnail(olio.image).get_file())
+		kuva.set_from_file(olio.thumbnail.get_file())
 		jako.add(kuva)
 		label = Gtk.Label(olio.name)
 		jako.add(label)
