+++ nn_parse.py
@@ -3,7 +3,7 @@
 from lxml import html
 import requests
 import re
-from os.path import expanduser
+from os.path import expanduser,isfile
 home = expanduser("~")	#multi os support
 import configparser
 import youtube_dl
@@ -38,6 +38,42 @@ class login(requests.Session):
 			self.group_names.append(kanava.xpath('a/text()')[0])
 			
 mie = login()
+
+class Kuva(object):
+	folder = None
+
+	def __init__(self, url):
+		self.imgname = re.search('\d+[.]jpg', url)
+		self.url = url
+		
+	def get_file(self):
+		return self.folder + self.imgname
+		
+	def write_file(self):
+		if not isfile(self.folder+self.imgname):				
+			response=requests.get(self.url)
+			with open(self.folder+self.imgname, 'wb') as img:
+				for chunk in response:
+					img.write(chunk)
+		
+class Avatar(Kuva):
+	folder = "./avatars/"
+	
+	def __init__(self, url):
+		Kuva.__init__(self, url)
+		self.imgname = self.imgname.group(0)
+		self.write_file()
+
+class Thumbnail(Kuva):
+	folder = "./thumbnails/"
+	
+	def __init__(self, url):
+		Kuva.__init__(self, url)
+		if self.imgname == None:
+			self.imgname="video.gif"
+		else:
+			self.imgname=self.imgname.group(0)
+		self.write_file()
 
 class User:
 	def __init__(self, user_id):
@@ -83,7 +119,8 @@ class Comment_user(User):
 	def __init__(self, comment_box):
 		user_id = comment_box.xpath('td[@class="author"]/div[@class="username"]/a/@href')[0]
 		User.__init__(self, user_id)
-		self.avatar = comment_box.xpath('td[@class="author_photo"]//img/@src')[0]
+		avatar_url = comment_box.xpath('td[@class="author_photo"]//img/@src')[0]
+		self.avatar = Avatar(avatar_url)
 		self.name = comment_box.xpath('td[@class="author"]//b/text()')[0]
 		self.comment_data = " ".join(comment_box.xpath('td[@class="author"]/div[@class="usergroup"]//text()')).strip()
 		
@@ -234,6 +271,16 @@ class Media:
 			}
 		mie.post("https://naurunappula.com/comment.php", payload)
 
+	def change_media(self, direction):
+		url = "https://naurunappula.com/go.php"
+		payload = {
+			'link_id': self.id,
+			'c': '2',
+			'dir': direction
+		}
+		return mie.get(url, params=payload)
+
+
 class VideoElement(Media):
 	def __init__(self, gridlist):
 		self.title = gridlist.xpath('@title')[0]
@@ -243,7 +290,8 @@ class VideoElement(Media):
 			self.name = gridlist.xpath('text()')[0]
 		self.link = "https://naurunappula.com" + gridlist.xpath('@href')[0]
 		media_id = re.search('\/(\d+)\/', self.link).group(1)
-		self.image = gridlist.xpath('img/@src')[0]
+		thumbnail_url = gridlist.xpath('img/@src')[0]
+		self.thumbnail = Thumbnail(thumbnail_url)
 		Media.__init__(self, media_id)
 	
 class VideoPage(Media):
